remove old version files.
