+++ .babelrc
@@ -1,6 +1,11 @@
 {
   "presets": [
-    ["es2015", {"modules": false}],
+    [
+      "es2015",
+      {
+        "modules": false
+      }
+    ],
     "stage-0"
   ]
 }
