+++ controllers/common.js
